requirements.txt
asyncpg~=0.23.0
py-cord
tbapi~=1.3.1b4
aiotba~=0.0.2.post1
googlemaps~=4.4.2
datetime~=4.3
geopy~=2.0.0
fuzzywuzzy~=0.18.0
python-Levenshtein
python-dateutil~=2.7.5
sentry-sdk
wavelink~=0.9.6
discord-ext-menus
rstcloth~=0.3.1
discord-py-slash-command
aiohttp~=3.6.2
humanize~=3.8.0
ChatterBot~=1.0.5